+++ packages/implementations/react/src/atomic/Chip/CHANGELOG.md
@@ -1,10 +1,13 @@
 # Changelog
 
-Description...
+Chip is a small component used to make selections, filter content, or trigger actions.
 
-## Date Version
+## 2024-04-16 `1.0.0`
 
 #### 🎉 New features
+
+- Added `children`, `pill`, `disabled`,`closeable`,`avatar`, `selected`and `icon` properties to the Component. ([#19](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/19) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
+- Added stories on Component. ([#19](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/19) by [@mauricio.barbosa](https://git.rarolabs.com.br/mauricio.barbosa))
 
 <!-- #### 🛠 Breaking changes -->